requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://pypi.nvidia.com

aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
asgiref==3.11.1
attrs==23.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/ultralytics/CLIP.git@88ade288431a46233f1556d1e141901b3ef0a36b
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.8
cuda-python==12.5.0
cvat-sdk==2.9.2
cycler==0.12.1
datumaro==1.5.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
Django==5.2.12
django-allauth==65.15.0
django-cors-headers==4.3.1
django-filter==25.2
django-import-export==3.3.4
djangorestframework==3.16.1
djangorestframework-api-key==3.1.0
drf-yasg==1.21.15
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.3.1
future==0.18.3
genicam==1.4.0
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.0.3
grpcio==1.64.1
gunicorn==25.1.0
h5py==3.11.0
harvesters==1.4.2
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.6
imageio-ffmpeg==0.6.0
inflection==0.5.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.4.2
json-stream==2.3.2
json-stream-rs-tokenizer==0.4.26
keyboard==0.13.5
kiwisolver==1.4.5
lap==0.5.13
lxml==5.2.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
newrelic==12.0.0
nibabel==5.2.1
numpy==1.26.2
odfpy==1.4.1
onnx==1.16.1
onnx-graphsurgeon==0.5.0
onnxruntime-gpu==1.18.1
onnxsim==0.4.36
onnxslim==0.1.88
opencv-python==4.8.1.78
openpyxl==3.1.2
openvino==2023.0.0
orjson==3.10.5
ovmsclient==2023.1
packaging==23.2
pandas==2.1.4
pefile==2023.2.7
Pillow==10.1.0
platformdirs==4.1.0
polars==1.38.1
polars-runtime-32==1.38.1 ; sys_platform == "win32"
protobuf==3.20.3
psutil==5.9.6
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pycocotools==2.0.8
pycparser==2.22
pyDeprecate==0.6.0
pyemd==1.0.0
Pygments==2.18.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
pyreadline3==3.4.1 ; sys_platform == "win32"
pyserial==3.5
python-dateutil==2.8.2
python-rapidjson==1.18
pytz==2023.3.post1
pywin32-ctypes==0.2.2 ; sys_platform == "win32"
PyYAML==6.0.1
regex==2026.1.15
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.4
six==1.16.0
sqlparse==0.4.4
supervision==0.27.0.post1
sympy==1.14.0
tablib==3.5.0
tabulate==0.9.0
tbb==2021.11.0
tensorboardX==2.6.2.2
tensorrt==10.0.1
tensorrt_dispatch==10.0.1
tensorrt_lean==10.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tinydb==4.8.0
tokenizers==0.19.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.1
tritonclient==2.47.0
tuspy==0.2.5
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.4.14
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.1.0
wcwidth==0.6.0
Werkzeug==3.1.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
